refactor: #13 결제처 repository 수정

+++ src/main/java/com/t3t/bookstoreapi/payment/repository/PaymentProviderRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PaymentProviderRepository extends JpaRepository<PaymentProvider, Long> {
+public interface PaymentProviderRepository extends JpaRepository<PaymentProvider, Integer> {
     PaymentProvider findByPaymentProviderName(String paymentProviderName);
 }
